+++ Backend/test_recommended_jobs.py
@@ -2,7 +2,7 @@ import unittest
 from unittest import mock
 from unittest.mock import Mock
 
-import recommend_jobs
+from Backend import recommend_jobs
 
 _range = 2
 interests_range = 1
@@ -56,7 +56,7 @@ class MyTestCase(unittest.TestCase):
         mocked_request.assert_called_with('https://api.lmiforall.org.uk/api/v1/o-net/soc2onet/1', verify=False)
         self.assertEqual(expected_data, actual_data)
 
-    @mock.patch('recommend_jobs.skill_sort')
+    @mock.patch('Backend.recommend_jobs.skill_sort')
     @mock.patch('requests.get')
     @mock.patch('requests.models.Response.json')
     def test_onet_skills(self, mocked_json, mocked_request, mocked_sort):
@@ -76,7 +76,7 @@ class MyTestCase(unittest.TestCase):
                                         {"id": "2.A.1.b", "name": "Active Listening", "value": 3.62}], skill_list)
         self.assertEqual(expected_data, actual_data)
 
-    @mock.patch('common.api_call')
+    @mock.patch('Backend.common.api_call')
     def test_onet_interests(self, mock_api):
         onet = '1'
         expected_data = [{'id': '1.B.1.a', 'name': 'Realistic', 'value': 1.33}]
@@ -148,12 +148,12 @@ class MyTestCase(unittest.TestCase):
         self.assertEqual(expected_data, actual_data)
 
     @mock.patch('builtins.input', side_effect=['engineer', 'No', 'engineer'])
-    @mock.patch('recommend_jobs.get_soc_code')
-    @mock.patch('recommend_jobs.soc_to_onet')
-    @mock.patch('recommend_jobs.onet_skills')
-    @mock.patch('recommend_jobs.onet_interests')
-    @mock.patch('recommend_jobs.reverse_search')
-    @mock.patch('recommend_jobs.find_job')
+    @mock.patch('Backend.recommend_jobs.get_soc_code')
+    @mock.patch('Backend.recommend_jobs.soc_to_onet')
+    @mock.patch('Backend.recommend_jobs.onet_skills')
+    @mock.patch('Backend.recommend_jobs.onet_interests')
+    @mock.patch('Backend.recommend_jobs.reverse_search')
+    @mock.patch('Backend.recommend_jobs.find_job')
     def test_run(self, mocked_job, mocked_search, mocked_onet_interests,
                  mocked_onet_skills, mocked_onet, mocked_soc, Mock):
         expected_data = ['plumber']
