+++ frontEnd/Main.html
@@ -1,0 +1,8 @@
+<h1>Government website for selling user data</h1>
+<form action="/action_page.php">
+  <label for="prev">Previous job:</label>
+  <input type="text" id="prev" name="prev"><br><br>
+  <label for="preferred">Preferred job:</label>
+  <input type="text" id="preferred" name="preferred"><br><br>
+  <input type="submit" value="Submit">
+</form>
